requirements.txt
streamlit>=1.28.0
google-generativeai>=0.3.0
pdf2image>=1.16.3
python-dotenv>=1.0.0
Pillow>=10.0.0